# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Trender Case API",
      "version": "1.0.0",
      "description": "API documentation for Trender Case"
    },
    "servers": [
      {
        "url": "http://localhost:3000/api"
      },
      {
        "url": "https://randomnumbersaver.azurewebsites.net/api"
      }
    ],
    "paths": {
      "/login": {
        "post": {
          "summary": "Login",
          "description": "Log in a user and return a JWT token",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully logged in",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid credentials",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Error logging in",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/register": {
        "post": {
          "summary": "Register",
          "description": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User registered successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Error registering user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Register failed",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/randomNumberSaver": {
        "post": {
          "summary": "Save Random Numbers",
          "description": "Save random numbers to the database",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "randomNumber1": {
                      "type": "integer",
                      "example": 42
                    },
                    "randomNumber2": {
                      "type": "integer",
                      "example": 84
                    }
                  },
                  "required": ["randomNumber1", "randomNumber2"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Numbers successfully saved to Database!",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Numbers stored successfully"
                      },
                      "status": {
                        "type": "integer",
                        "example": 201
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Invalid input"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Authorization token missing or invalid",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Authorization token missing"
                      },
                      "status": {
                        "type": "integer",
                        "example": 401
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Error storing numbers",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Error storing numbers"
                      },
                      "error": {
                        "type": "string",
                        "example": "Detailed error message"
                      },
                      "status": {
                        "type": "integer",
                        "example": 500
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }
  